Enable combined 'geocode|establishment' type
The default query type is address, while the default API type is 'geocode|establishment'. The typings are preventing us from using the combined type, which returns both addresses and businesses.

I am proposing to change the typings so at least we can use this option. It would be even better if the default type was the same than the API choice.

See:
https://developers.google.com/places/web-service/autocomplete#place_types

    "The exception is that you can safely mix the `geocode` and `establishment` types, but note that this will have the same effect as specifying no types."